web-back/requirements.txt
Django
django-cors-headers
djangorestframework
Pillow
gunicorn
psycopg2-binary
django-debug-toolbar
django-cleanup
